### 2023-05-03 #5433 @editorialbot
[PRE REVIEW]: Snek5000: a new Python framework for Nek5000
**Submitting author:** @paugier (<a href="http://orcid.org/0000-0001-9481-4459">Pierre Augier</a>)
**Repository:** https://github.com/snek5000/snek5000
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.0
