### 2023-05-06 6c7e65a
Reduce pre-commit autoupdate frequency
Monthly
