+++ requirements/main.txt
@@ -126,7 +126,7 @@ pyyaml==6.0
     #   yte
 qtpy==2.3.0
     # via fluiddyn
-requests==2.28.1
+requests==2.31.0
     # via snakemake
 reretry==0.11.1
     # via snakemake
