### 2023-06-08 #5433 @philipcardiff
Hello @paugier, I have suggested some minor English edits at https://github.com/snek5000/snek5000/pull/293.
