### 2023-06-08 #5433 @philipcardiff
Hello @stgeke, would you review this submission for JOSS? The submission under consideration is Snek5000: a new Python framework for Nek5000 (https://github.com/snek5000/snek5000). Your [nekpy](https://github.com/maxhutch/nekpy) repository is mentioned in the article.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected.

Philip
