### 2023-06-14 #5433 @ashwinvis
@philipcardiff Co-author of the article here. I could suggest a couple of potential reviewers. One who had enquired about Snek5000 in the past via email and another who submitted a bug report on a different open source project. I have met or worked with them professionally. I believe there is no conflict of interest. Can you confirm?
