### 2023-06-29 #5586 @joneuhauser
@philipcardiff I have opened a few issues in the repository when I reviewed it last week. I'm not using this project regularly, so only what I found from review-related testing (i.e. trying out the tutorials).

I recommend acceptance in JOSS.
