### 2023-07-03 9067a88
[pre-commit.ci] pre-commit autoupdate
updates:
- https://github.com/charliermarsh/ruff-pre-commit → https://github.com/astral-sh/ruff-pre-commit
- [github.com/astral-sh/ruff-pre-commit: v0.0.263 → v0.0.276](https://github.com/astral-sh/ruff-pre-commit/compare/v0.0.263...v0.0.276)
