+++ .pre-commit-config.yaml
@@ -16,8 +16,8 @@ repos:
   rev: 23.3.0
   hooks:
   - id: black
-- repo: https://github.com/charliermarsh/ruff-pre-commit
-  rev: v0.0.263
+- repo: https://github.com/astral-sh/ruff-pre-commit
+  rev: v0.0.276
   hooks:
   - id: ruff
 - repo: https://github.com/snakemake/snakefmt
