### 2023-07-04 e27b714
Fix example in README
