### 2023-07-04 8e1f2ac
Fix usage of old deprecated function get_configfile
