### 2023-07-05 bf0c195
Separate code snippet on history points
