### 2023-07-31 snek5000/snek5000#302 @maxhutch
Quickstart phill example fails with Nek500 v19
When running the phill example with Nek5000 v19 (which is the version downloadable from the official site [here](https://nek5000.mcs.anl.gov/)), the following error is reported:
```
/----------------------------------------------------------\\
|      _   __ ______ __ __  ______  ____   ____   ____     |
|     / | / // ____// //_/ / ____/ / __ \\/ __ \\/ __ \\   |
|    /  |/ // __/  / ,<   /___ \\ / / / // / / // / / /    |
|   / /|  // /___ / /| | ____/ / / /_/ // /_/ // /_/ /     |
|  /_/ |_//_____//_/ |_|/_____/  \\___/ \\___/ \\___/      |
|                                                          |
|----------------------------------------------------------|
|                                                          |
| COPYRIGHT (c) 2008-2019 UCHICAGO ARGONNE, LLC            |
| Version:  19.0                                           |
| Web:      https://nek5000.mcs.anl.gov                    |
|                                                          |
\\----------------------------------------------------------/

 Number of MPI ranks :           8

 Reading ./session_00/phill.par
 ERROR: Par file contains unknown key GENERAL:USERPARAM12

an error occured: dying ...
```

Running with the main version on github, [this version at the time of writing](https://github.com/Nek5000/Nek5000/tree/11e94661faa8991b5c6e6711daf89d1cf1aa3ed5), it does work, so this may be an implicit version dependency.
