### 2023-08-01 snek5000/snek5000#302 @ashwinvis
@maxhutch yes, that's correct. We used to support v19, but now we need a more recent version containing some enhancements and bugfixes (but not the master version). See the thread in #294 and the changes we did in #299.
