### 2023-08-01 b5a7780
README: link towards notes on incompatible Nek5000 versions
