### 2023-08-01 snek5000/snek5000#302 @paugier
Thanks @maxhutch for filling this issue. As @ashwinvis wrote, we already have some notes in the documentation mentioning this known incompatibility. In #304, I added a line in the README to mention these notes. I hope it's going to be clearer like that.
