### 2023-08-08 #5586 @philipcardiff
Hi @maxhutch, I see you mostly completed the reviewer checklist and the issue you raised (https://github.com/snek5000/snek5000/issues/302) was resolved.

If you have any questions/comments regarding the final checklist points (Substantial scholarly effort, Functionality, and Functionality documentation), please raise them here or as an issue in the repository.
