### 2023-08-16 #5586 @maxhutch
I recommend acceptance in JOSS as well.
