### 2023-08-20 281373f
Rm editorconfig, vim, activate,sh
