+++ setup.cfg
@@ -19,6 +19,7 @@ classifiers =
         Programming Language :: Python :: 3.10
         Topic :: Scientific/Engineering
 project-urls =
+    Sponsor = https://fluid.quest/pages/sponsors.html
     Bug Tracker = https://github.com/snek5000/snek5000/issues
     Documentation = https://snek5000.readthedocs.io
     Changelog = https://snek5000.readthedocs.io/en/latest/CHANGELOG.html
