+++ requirements/docs.txt
@@ -1,6 +1,6 @@
 #
-# This file is autogenerated by pip-compile with Python 3.10
-# by the following command:
+# This file is autogenerated by pip-compile with python 3.8
+# To update, run:
 #
 #    pip-compile --extra=docs --output-file=requirements/docs.txt --resolver=backtracking requirements/vcs_packages.in setup.cfg
 #
@@ -11,58 +11,63 @@
     #   snek5000-cbox
     #   snek5000-phill
     #   snek5000-tgv
-alabaster==0.7.12
+alabaster==0.7.13
     # via sphinx
-anyio==3.6.2
+anyio==3.7.1
     # via jupyter-server
 appdirs==1.4.4
     # via snakemake
-argon2-cffi==21.3.0
+argon2-cffi==23.1.0
     # via jupyter-server
 argon2-cffi-bindings==21.2.0
     # via argon2-cffi
+arrow==1.2.3
+    # via isoduration
 asttokens==2.2.1
     # via stack-data
-attrs==22.1.0
+attrs==23.1.0
     # via
     #   jsonschema
     #   jupyter-cache
     #   pymech
-babel==2.11.0
+    #   referencing
+babel==2.12.1
     # via sphinx
 backcall==0.2.0
     # via ipython
-beautifulsoup4==4.11.1
+beautifulsoup4==4.12.2
     # via
     #   furo
     #   nbconvert
-bleach==5.0.1
+bleach==6.0.0
     # via nbconvert
-breathe==4.34.0
+breathe==4.35.0
     # via snek5000 (setup.cfg)
-certifi==2022.9.24
+certifi==2023.7.22
     # via requests
 cffi==1.15.1
     # via argon2-cffi-bindings
-charset-normalizer==2.1.1
+charset-normalizer==3.2.0
     # via requests
-click==8.1.3
+click==8.1.7
     # via jupyter-cache
 colorama==0.4.6
     # via
     #   nbdime
     #   sphinx-autobuild
-configargparse==1.5.3
+comm==0.1.4
+    # via ipykernel
+configargparse==1.7
     # via snakemake
 connection-pool==0.0.3
     # via snakemake
-contourpy==1.0.6
+contourpy==1.1.0
     # via matplotlib
 cycler==0.11.0
     # via matplotlib
 datrie==0.8.2
     # via snakemake
-debugpy==1.6.4
+debugpy==1.6.7.post1
     # via ipykernel
 decorator==5.1.1
     # via ipython
@@ -77,66 +82,82 @@ docutils==0.17.1
     #   snakemake
     #   snek5000 (setup.cfg)
     #   sphinx
-dpath==2.1.2
+dpath==2.1.6
     # via yte
-entrypoints==0.4
-    # via jupyter-client
+exceptiongroup==1.1.3
+    # via anyio
 executing==1.2.0
     # via stack-data
-fastjsonschema==2.16.2
+fastjsonschema==2.18.0
     # via nbformat
-filelock==3.8.1
+filelock==3.12.2
     # via
     #   snek5000
     #   snek5000 (setup.cfg)
-fluiddyn==0.5.1
+fluiddyn==0.5.2
     # via
     #   fluidsim-core
     #   snek5000
     #   snek5000 (setup.cfg)
-fluidsim-core==0.7.2
+fluidsim-core==0.7.3
     # via
     #   snek5000
     #   snek5000 (setup.cfg)
-fonttools==4.38.0
+fonttools==4.42.1
     # via matplotlib
-furo==2022.9.29
+fqdn==1.5.1
+    # via jsonschema
+furo==2023.3.27
     # via snek5000 (setup.cfg)
 gitdb==4.0.10
[…]
