### 2023-08-23 502c350
Replace project-urls with project_urls
Warning from setuptools

skip-checks: true
