+++ setup.cfg
@@ -18,7 +18,7 @@ classifiers =
         Programming Language :: Python :: 3.9
         Programming Language :: Python :: 3.10
         Topic :: Scientific/Engineering
-project-urls =
+project_urls =
     Sponsor = https://fluid.quest/pages/sponsors.html
     Bug Tracker = https://github.com/snek5000/snek5000/issues
     Documentation = https://snek5000.readthedocs.io
