### 2023-08-24 #5586 @ashwinvis
@philipcardiff
The new version is v0.9.2 (or git tag 0.9.2)
I have also archived it to Zenodo (https://zenodo.org/record/8278552) (doi: 10.5281/zenodo.8278552)
