### 2023-08-24 5a75ce2
Update JOSS paper
Minor typographic edits to JOSS paper
