### 2023-08-24 c304414
Update paper.bib
Capitalizes some words in reference titles
