+++ docs/joss_paper/paper.bib
@@ -13,7 +13,7 @@ year = {2019},
 publisher = {Ubiquity Press,  Ltd.},
 volume = {7},
 author = {Pierre Augier and Ashwin Vishnu Mohanan and Cyrille Bonamy},
-title = {{FluidDyn}: A Python Open-Source Framework for Research and Teaching in Fluid Dynamics
+title = {{FluidDyn}: A {Python} Open-Source Framework for Research and Teaching in Fluid Dynamics
     by Simulations,  Experiments and Data Processing},
 journal = {Journal of Open Research Software}
 }
@@ -42,7 +42,7 @@ year = {2022}
 
 @article{snakemake,
 author = { Mölder, F and Jablonski, KP and Letcher, B and Hall, MB and Tomkins-Tinch, CH and Sochat, V and Forster, J and Lee, S and Twardziok, SO and Kanitz, A and Wilm, A and Holtgrewe, M and Rahmann, S and Nahnsen, S and Köster, J},
-title = {Sustainable data analysis with Snakemake [version 1; peer review: 1 approved, 1 approved with reservations]
+title = {Sustainable data analysis with {Snakemake} [version 1; peer review: 1 approved, 1 approved with reservations]
 },
 journal = {F1000Research},
 volume = {10},
@@ -53,7 +53,7 @@ doi = {10.12688/f1000research.29032.1}
 
 @article{matplotlib,
   Author    = {Hunter, J. D.},
-  Title     = {Matplotlib: A 2D graphics environment},
+  Title     = {Matplotlib: A {2D} graphics environment},
   Journal   = {Computing in Science \& Engineering},
   Volume    = {9},
   Number    = {3},
@@ -79,7 +79,7 @@ doi = {10.12688/f1000research.29032.1}
 }
 
 @article{nekrs,
-  title={NekRS, a GPU-accelerated spectral element Navier--Stokes solver},
+  title={NekRS, a {GPU}-accelerated spectral element {Navier}--{Stokes} solver},
   author={Fischer, Paul and Kerkemeier, Stefan and Min, Misun and Lan, Yu-Hsiang and Phillips, Malachi and Rathnayake, Thilina and Merzari, Elia and Tomboulides, Ananias and Karakus, Ali and Chalmers, Noel and others},
   journal={Parallel Computing},
   volume={114},
@@ -91,7 +91,7 @@ doi = {10.12688/f1000research.29032.1}
 
 @inproceedings{nek5000_scaling,
   author = {Offermans, Nicolas and Marin, Oana and Schanen, Michel and Gong, Jing and Fischer, Paul and Schlatter, Philipp and Obabko, Aleks and Peplinski, Adam and Hutchinson, Maxwell and Merzari, Elia},
-  title = {On the Strong Scaling of the Spectral Element Solver Nek5000 on Petascale Systems},
+  title = {On the Strong Scaling of the Spectral Element Solver {Nek5000} on Petascale Systems},
   year = {2016},
   isbn = {9781450341226},
   publisher = {Association for Computing Machinery},
