### 2023-08-24 #5586 @kyleniemeyer
Hi @paugier, I made a few small edits to the paper. Can you review and merge these? https://github.com/snek5000/snek5000/pull/308
