### 2023-08-24 1365d55
Some more titlecase and lowercase changes to references
