+++ docs/joss_paper/paper.bib
@@ -25,8 +25,8 @@ publisher = {Ubiquity Press,  Ltd.},
 volume = {7},
 author = {Mohanan, Ashwin Vishnu and Bonamy, Cyrille and Linares, Miguel
     Calpe and Augier, Pierre},
-title = {{FluidSim}: {Modular}, {Object}-{Oriented} {Python} {Package} for
-    {High}-{Performance} {CFD} {Simulations}},
+title = {{FluidSim}: modular, object-oriented {Python} package for
+    high-performance {CFD} simulations},
 journal = {Journal of Open Research Software}
 }
 
@@ -114,7 +114,7 @@ doi = {10.12688/f1000research.29032.1}
       Adam and Jansson, Niclas and Podobas, Artur and Jocksch, Andreas and Yao,
       Jie and Hussain, Fazle and Markidis, Stefano and Karlsson, Matts and
           Pleiter, Dirk and Laure, Erwin and Schlatter, Philipp},
-  title = {Strong Scaling of OpenACC Enabled Nek5000 on Several GPU Based HPC Systems},
+  title = {Strong Scaling of {OpenACC} Enabled {Nek5000} on Several {GPU} Based {HPC} Systems},
   year = {2022},
   isbn = {9781450384988},
   publisher = {Association for Computing Machinery},
@@ -151,7 +151,7 @@ doi = {10.12688/f1000research.29032.1}
 }
 
 @article{Khoubani2023vertical,
-  title={Vertical convection regimes in a rectangular cavity: Prandtl and aspect ratio dependance},
+  title={Vertical convection regimes in a rectangular cavity: {Prandtl} and aspect ratio dependance},
   author={Arman Khoubani and Ashwin Vishnu Mohanan and Pierre Augier and Jan-Bert Flór},
   journal={Submitted to the Journal of Fluid Mechanics},
   year={2023},
