### 2023-08-24 897c360
Article accepted in JOSS! Update DOI, CFF & badges [skip ci]
