+++ CITATION.cff
@@ -21,7 +21,11 @@ authors:
     orcid: https://orcid.org/0000-0001-9481-4459
 
 version: 0.9.2
-date-released: 2023-08-13
-doi: 10.5281/zenodo.7399621
+date-released: 2023-08-24
+doi: 10.21105/joss.05586
 license: BSD-3-Clause
 repository-code: "https://github.com/snek5000/snek5000"
+identifiers:
+  - type: doi
+    value: 10.5281/zenodo.7399621
+    description: The concept DOI for the collection containing all versions of Snek5000 source code
