### 2023-08-24 2c55979
Title and abstract as in the JOSS paper
