### 2023-08-25 #5586 @paugier
Thanks a lot @kyleniemeyer, @joneuhauser, @maxhutch and @philipcardiff for your work on JOSS and on this paper. The review process in JOSS is really a pleasant experience!

@ashwinvis and @akhoubani, we can be happy with this achievement. This was much harder than using Nek5000 with quick and dirty scripting but Snek5000 is a real added value in the long term. @akhoubani this is also a very good news for your PhD thesis!
